ids/requirements.txt
# IDS Service Dependencies
flask>=2.3.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tensorflow-cpu>=2.13.0
joblib>=1.3.0
